Fix type of AC field tempo Include : in the replace operation to avoid replacing partially matched field names (e.g. ac_tempo in ac_tempo_confidence)
https://github.com/MTG/freesound/pull/1238

+++ freesound/settings.py
@@ -202,7 +202,7 @@ AUDIOCOMMONS_INCLUDED_DESCRIPTOR_NAMES_TYPES = \
      ('tonality', str),
      ('tonality_confidence', float),
      ('loop', bool),
-     ('tempo', float),
+     ('tempo', int),
      ('tempo_confidence', float),
      ('note_midi', int),
      ('note_name', str),
